+++ lib/rubocop/cop/style/multiline_operation_indentation.rb
@@ -114,10 +114,14 @@ module RuboCop
           if operator?(method_name) && args.any?
             args.first.loc.expression
           elsif send_node.loc.dot &&
+                send_node.loc.selector &&
                 send_node.loc.dot.line == send_node.loc.selector.line
             send_node.loc.dot.join(send_node.loc.selector)
-          else
+          elsif send_node.loc.selector
             send_node.loc.selector
+          elsif send_node.loc.dot.line == send_node.loc.begin.line
+            # lambda.(args)
+            send_node.loc.dot.join(send_node.loc.begin)
           end
         end